Table will only sort on load if defaultSort has been defined.
It will load in the order the data has been received from the server, otherwise -- headers can still be clicked for sorting.  This will prevent issues in certain situations where sort order from the server should be preserved and not overridden on table load.  Should also remove conflicts when Drag and Drop is implemented.